Dev back

+++ back/auth/internal/app/tokenizer/tokenizer_interface.go
@@ -3,9 +3,9 @@ package tokenizer
 import "time"
 
 type Tokenizer interface {
-	CreateToken(employeeId uint, companyId uint) (string, error)
+	CreateToken(employeeId uint64, companyId uint64) (string, error)
 	CheckExpiration(token string) (bool, error)
-	DecryptToken(token string) (uint, uint, error)
+	DecryptToken(token string) (uint64, uint64, error)
 }
 
 func New(
