+++ back/auth/internal/model/employee.go
@@ -3,6 +3,6 @@ package model
 type Employee struct {
 	Email      string
 	Password   string
-	EmployeeId uint
-	CompanyId  uint
+	EmployeeId uint64
+	CompanyId  uint64
 }
