+++ back/brm-core/internal/model/company.go
@@ -3,11 +3,11 @@ package model
 import "time"
 
 type Company struct {
-	Id           uint
+	Id           uint64
 	Name         string
 	Description  string
-	Industry     uint
-	OwnerId      uint
+	Industry     uint64
+	OwnerId      uint64
 	Rating       float64
 	CreationDate time.Time
 	IsDeleted    bool
@@ -16,6 +16,6 @@ type Company struct {
 type UpdateCompany struct {
 	Name        string
 	Description string
-	Industry    uint
-	OwnerId     uint
+	Industry    uint64
+	OwnerId     uint64
 }
