+++ back/brm-core/internal/repo/repo_interface.go
@@ -20,27 +20,27 @@ func New(conn *pgx.Conn) CoreRepo {
 }
 
 type CompanyRepo interface {
-	GetCompany(ctx context.Context, id uint) (model.Company, error)
+	GetCompany(ctx context.Context, id uint64) (model.Company, error)
 	CreateCompanyAndOwner(ctx context.Context, company model.Company, owner model.Employee) (model.Company, model.Employee, error)
-	UpdateCompany(ctx context.Context, companyId uint, upd model.UpdateCompany) (model.Company, error)
-	DeleteCompany(ctx context.Context, companyId uint) error
+	UpdateCompany(ctx context.Context, companyId uint64, upd model.UpdateCompany) (model.Company, error)
+	DeleteCompany(ctx context.Context, companyId uint64) error
 }
 
 type EmployeeRepo interface {
 	CreateEmployee(ctx context.Context, employee model.Employee) (model.Employee, error)
-	UpdateEmployee(ctx context.Context, employeeId uint, upd model.UpdateEmployee) (model.Employee, error)
-	DeleteEmployee(ctx context.Context, employeeId uint) error
+	UpdateEmployee(ctx context.Context, employeeId uint64, upd model.UpdateEmployee) (model.Employee, error)
+	DeleteEmployee(ctx context.Context, employeeId uint64) error
 
-	GetCompanyEmployees(ctx context.Context, companyId uint, filter model.FilterEmployee) ([]model.Employee, error)
-	GetEmployeeByName(ctx context.Context, companyId uint, ebn model.EmployeeByName) ([]model.Employee, error)
-	GetEmployeeById(ctx context.Context, employeeId uint) (model.Employee, error)
+	GetCompanyEmployees(ctx context.Context, companyId uint64, filter model.FilterEmployee) ([]model.Employee, error)
+	GetEmployeeByName(ctx context.Context, companyId uint64, ebn model.EmployeeByName) ([]model.Employee, error)
+	GetEmployeeById(ctx context.Context, employeeId uint64) (model.Employee, error)
 }
 
 type ContactRepo interface {
 	CreateContact(ctx context.Context, contact model.Contact) (model.Contact, error)
-	UpdateContact(ctx context.Context, ownerId uint, contactId uint, upd model.UpdateContact) (model.Contact, error)
-	DeleteContact(ctx context.Context, ownerId uint, contactId uint) error
+	UpdateContact(ctx context.Context, ownerId uint64, contactId uint64, upd model.UpdateContact) (model.Contact, error)
+	DeleteContact(ctx context.Context, ownerId uint64, contactId uint64) error
 
-	GetContacts(ctx context.Context, ownerId uint, pagination model.GetContacts) ([]model.Contact, error)
-	GetContactById(ctx context.Context, ownerId uint, contactId uint) (model.Contact, error)
+	GetContacts(ctx context.Context, ownerId uint64, pagination model.GetContacts) ([]model.Contact, error)
+	GetContactById(ctx context.Context, ownerId uint64, contactId uint64) (model.Contact, error)
 }
