+++ back/registration/internal/ports/httpserver/responses.go
@@ -51,8 +51,8 @@ type companyAndOwnerData struct {
 }
 
 type ownerData struct {
-	Id           uint   `json:"id"`
-	CompanyId    uint   `json:"company_id"`
+	Id           uint64 `json:"id"`
+	CompanyId    uint64 `json:"company_id"`
 	FirstName    string `json:"first_name"`
 	SecondName   string `json:"second_name"`
 	Email        string `json:"email"`
@@ -63,11 +63,11 @@ type ownerData struct {
 }
 
 type companyData struct {
-	Id           uint    `json:"id"`
+	Id           uint64  `json:"id"`
 	Name         string  `json:"name"`
 	Description  string  `json:"description"`
-	Industry     uint    `json:"industry"`
-	OwnerId      uint    `json:"owner_id"`
+	Industry     uint64  `json:"industry"`
+	OwnerId      uint64  `json:"owner_id"`
 	Rating       float64 `json:"rating"`
 	CreationDate int64   `json:"creation_date"`
 	IsDeleted    bool    `json:"is_deleted"`
