+++ back/transport-api/internal/app/contact.go
@@ -5,22 +5,22 @@ import (
 	"transport-api/internal/model/core"
 )
 
-func (a *appImpl) CreateContact(ctx context.Context, ownerId uint, employeeId uint) (core.Contact, error) {
+func (a *appImpl) CreateContact(ctx context.Context, ownerId uint64, employeeId uint64) (core.Contact, error) {
 	return a.core.CreateContact(ctx, ownerId, employeeId)
 }
 
-func (a *appImpl) UpdateContact(ctx context.Context, ownerId uint, contactId uint, upd core.UpdateContact) (core.Contact, error) {
+func (a *appImpl) UpdateContact(ctx context.Context, ownerId uint64, contactId uint64, upd core.UpdateContact) (core.Contact, error) {
 	return a.core.UpdateContact(ctx, ownerId, contactId, upd)
 }
 
-func (a *appImpl) DeleteContact(ctx context.Context, ownerId uint, contactId uint) error {
+func (a *appImpl) DeleteContact(ctx context.Context, ownerId uint64, contactId uint64) error {
 	return a.core.DeleteContact(ctx, ownerId, contactId)
 }
 
-func (a *appImpl) GetContacts(ctx context.Context, ownerId uint, pagination core.GetContacts) ([]core.Contact, error) {
+func (a *appImpl) GetContacts(ctx context.Context, ownerId uint64, pagination core.GetContacts) ([]core.Contact, error) {
 	return a.core.GetContacts(ctx, ownerId, pagination)
 }
 
-func (a *appImpl) GetContactById(ctx context.Context, ownerId uint, contactId uint) (core.Contact, error) {
+func (a *appImpl) GetContactById(ctx context.Context, ownerId uint64, contactId uint64) (core.Contact, error) {
 	return a.core.GetContactById(ctx, ownerId, contactId)
 }
