+++ back/transport-api/internal/ports/httpserver/core/companies/presenters.go
@@ -3,6 +3,6 @@ package companies
 type updateCompanyRequest struct {
 	Name        string `json:"name"`
 	Description string `json:"description"`
-	Industry    uint   `json:"industry"`
-	OwnerId     uint   `json:"owner_id"`
+	Industry    uint64 `json:"industry"`
+	OwnerId     uint64 `json:"owner_id"`
 }
