+++ back/transport-api/internal/ports/httpserver/core/contacts/presenters.go
@@ -2,7 +2,7 @@ package contacts
 
 type addContactRequest struct {
 	// EmployeeId это id сотрудника, которого добавляют в контакты
-	EmployeeId int `json:"employee_id"`
+	EmployeeId uint64 `json:"employee_id"`
 }
 
 type updateContactRequest struct {
